+++ src/refactorings/extract-variable/extract-variable.ts
@@ -182,7 +182,8 @@ function isExtractableContext(node: ast.Node): boolean {
     ast.isClassProperty(node) ||
     ast.isIfStatement(node) ||
     ast.isWhileStatement(node) ||
-    ast.isSwitchCase(node)
+    ast.isSwitchCase(node) ||
+    ast.isJSXExpressionContainer(node)
   );
 }